pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "logos-foundry"
version = "0.1.0"
description = "LOGOS Foundry - Canonical source of truth for Project LOGOS specifications, ontology, and infrastructure"
readme = "README.md"
authors = ["Project LOGOS Contributors"]
license = "MIT"
keywords = [
    "logos",
    "hybrid-causal-graph",
    "hcg",
    "neo4j",
    "milvus",
    "ontology",
    "shacl",
    "cognitive-architecture"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
packages = [
    { include = "logos_tools" },
    { include = "logos_hcg" },
    { include = "logos_observability" },
    { include = "logos_persona" },
    { include = "logos_cwm_e" },
    { include = "logos_perception" },
    { include = "logos_sophia" },
    { include = "logos_test_utils" },
    { include = "planner_stub" }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.0.0"

# Test utilities are available but optional (dev/test only)
[tool.poetry.group.test.dependencies]
rdflib = "^7.0.0"
pyshacl = "^0.25.0"
neo4j = "^6.0.0"
pymilvus = "^2.3.0"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-instrumentation-fastapi = "^0.41b0"
opentelemetry-exporter-otlp-proto-grpc = "^1.20.0"
numpy = "^1.24.0"
requests = "^2.32.5"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.0"
black = "^24.0.0"
mypy = "^1.7.0"
httpx = "^0.27.0"
requests = "^2.31.0"
types-PyYAML = "^6.0.12.12"
types-requests = "^2.31.0"

[tool.poetry.scripts]
logos-generate-issues = "logos_tools.generate_issues:main"
logos-create-issues-by-epoch = "logos_tools.create_issues_by_epoch:main"
render-test-stacks = "infra.scripts.render_test_stacks:main"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["logos_tools", "logos_hcg", "logos_observability", "logos_persona", "logos_cwm_e"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 100
target-version = ["py310"]
extend-exclude = '''
/(
    sdk
  | sdk-web
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [
    "sdk",
    "sdk-web",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = ["sdk", "sdk-web"]
